refactor: 이미지 업로드, 읽는 기능하는 useImageFileHandler 훅 구현 (#23)
